Final Project/Flask/requirements.txt
pandas==0.22.0
numpy==1.13.3
pickleshare==0.7.4
urllib3==1.22
sklearn
shutil
logging
boto.s3
os
zipfile
flask
functions
werkzeug
json
sys